+++ lib/objects/product_related/psu_related/psu.dart
@@ -39,8 +39,8 @@ class PSU extends Product {
   void updateProduct({
     String? productID,
     String? productName,
-    double? importPrice,
-    double? sellingPrice,
+    int? importPrice,
+    int? sellingPrice,
     double? discount,
     DateTime? release,
     int? sales,
