+++ lib/objects/product_related/ram_related/ram.dart
@@ -39,8 +39,8 @@ class RAM extends Product {
   void updateProduct({
     String? productID,
     String? productName,
-    double? importPrice,
-    double? sellingPrice,
+    int? importPrice,
+    int? sellingPrice,
     double? discount,
     DateTime? release,
     int? sales,
